Extract topic store state
